+++ src/features/settings/components/display/DisplaySettingsSection.tsx
@@ -1,4 +1,4 @@
-import { FaMoon } from "react-icons/fa6";
+import { FaBrush } from "react-icons/fa6";
 import { useTheme } from "@features/settings/hooks/useTheme";
 import { ThemeToggle } from "./ThemeToggle";
 import { SettingsCard } from "../SettingsCard";
@@ -9,7 +9,7 @@ export function DisplaySettingsSection() {
   return (
     <div className="mx-auto max-w-lg w-full flex flex-col items-center">
       <h2 className="text-2xl font-bold mb-6 self-start">Display Settings</h2>
-      <SettingsCard title="Theme" icon={<FaMoon />}>
+      <SettingsCard title="Theme" icon={<FaBrush />}>
         <div className="flex flex-col gap-2 w-full">
           <div className="flex items-center gap-4 mb-2">
             <ThemeToggle theme={theme} toggleTheme={toggleTheme} />